remove lint warnings, update tests and rename file
