+++ packages/cli/babel/index.js
@@ -1,3 +1,5 @@
+// Is this file used/required? If not, can it be removed?
+
 var isProd = (process.env.BABEL_ENV || process.env.NODE_ENV) === 'production';
 /**
  * test env detection is used to default mode for
@@ -21,7 +23,7 @@ var defaultBrowserListDev = [
 	'ios_saf >= 10.3',
 	'edge >= 16',
 	'opera >= 48',
-	'samsung >= 8.2'
+	'samsung >= 8.2',
 ];
 
 // preact-cli babel configs
