+++ packages/cli/tests/build.test.js
@@ -182,8 +182,10 @@ describe('preact build', () => {
 			`${dir}/build/${bundleFile}`,
 			'utf8'
 		);
+
 		// "Hello World!" should replace 'process.env.PREACT_APP_MY_VARIABLE'
-		expect(transpiledChunk.includes('console.log("Hello World!")')).toBe(true);
+		expect(transpiledChunk.includes('Hello World!')).toBe(true);
+		expect(transpiledChunk.includes('PREACT_APP_MY_VARIABLE')).toBe(false);
 	});
 
 	it('should respect `publicPath` value', async () => {
