+++ packages/cli/tests/lib/output.js
@@ -10,7 +10,7 @@ async function tmpDir() {
 	let str = Math.random()
 		.toString(36)
 		.replace(/[^a-z]+/g, '')
-		.substr(0, 12);
+		.substring(0, 12);
 	await mkdir(output, { recursive: true });
 	return resolve(output, str);
 }
