+++ packages/cli/tests/subjects/custom-dotenv/index.js
@@ -1,1 +1,1 @@
-console.log(process.env.PREACT_APP_MY_VARIABLE);
+export const MY_VARIABLE = process.env.PREACT_APP_MY_VARIABLE;
